requirements.txt
Django==5.1.7
djangorestframework==3.12.0
djangorestframework_simplejwt==4.8.0
django-cors-headers==3.7.0
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg2==2.9.10
whitenoise==6.9.0
python-decouple==3.8